+++ pwa/firebase-config.js
@@ -1,8 +1,8 @@
-window.I_EXPENSE_FIREBASE_CONFIG = window.I_EXPENSE_FIREBASE_CONFIG || {
-  apiKey: "",
-  authDomain: "",
-  projectId: "",
-  appId: "",
-  messagingSenderId: "",
-  storageBucket: ""
+window.I_EXPENSE_FIREBASE_CONFIG = {
+  apiKey: "AIzaSyBflzaANT9BnCbMfBVbjvxyM8KnirJBAvo",
+  authDomain: "i-expense-pwa.firebaseapp.com",
+  projectId: "i-expense-pwa",
+  appId: "1:160747680735:web:1ce0be9463eb033f318a6a",
+  messagingSenderId: "160747680735",
+  storageBucket: "i-expense-pwa.firebasestorage.app"
 };